EDA_Tool_AWS/pyproject.toml
[project]
name = "eda-tool"
version = "1.0.0"
description = "A comprehensive web-based Exploratory Data Analysis tool for Data Science teams"
authors = [
    {name = "Data Science Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "openpyxl>=3.1.0",
    "xlrd>=2.0.0",
    "plotly>=5.17.0",
    "scipy>=1.11.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.12.0",
    "jinja2>=3.1.0",
    "werkzeug>=3.0.0",
    "python-dateutil>=2.8.0",
    "reportlab>=4.0.0",
    "weasyprint>=60.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.1.1",
]
